More encoder and DCC fixes.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@5220 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_encoder.c
@@ -72,7 +72,7 @@ mx_encoder_get_pointers( MX_RECORD *encoder_record,
 }
 
 MX_EXPORT mx_status_type
-mx_encoder_read( MX_RECORD *encoder_record, long *value )
+mx_encoder_read( MX_RECORD *encoder_record, double *value )
 {
 	static const char fname[] = "mx_encoder_read()";
 
@@ -103,7 +103,7 @@ mx_encoder_read( MX_RECORD *encoder_record, long *value )
 }
 
 MX_EXPORT mx_status_type
-mx_encoder_write( MX_RECORD *encoder_record, long value )
+mx_encoder_write( MX_RECORD *encoder_record, double value )
 {
 	static const char fname[] = "mx_encoder_write()";
